Revert "Version"
This reverts commit 302c46daa596eaec9d8baf2dfd95606a4bc24888.
