+++ .spyproject/config/workspace.ini
@@ -6,5 +6,5 @@ save_non_project_files = False
 
 [main]
 version = 0.2.0
-recent_files = ['main.py', 'error_classes.py']
+recent_files = []
