+++ error_classes.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 class AssemblerError(Exception):
     pass
 
@@ -26,21 +28,3 @@ class AssemblerMemoryError(AssemblerError):
 
     def __str__(self):
         return f"QCPU instruction memory limit exceeded: {self.reason}"
-
-
-class AssemblerTypeError(AssemblerError):
-    def __init__(self, line, reason):
-        self.line = line
-        self.reason = reason
-
-    def __str__(self):
-        return f"Type error on line {self.line}: {self.reason}"
-
-
-class AssemblerInternalError(AssemblerError):
-    def __init__(self, line, reason):
-        self.line = line
-        self.reason = reason
-
-    def __str__(self):
-        return f"Internal error on line {self.line}: {self.reason}"
