--- output/Bresenhams-1.txt
@@ -1,28 +1,0 @@
-00101000
-01000110
-00100000
-00101001
-01000001
-00101001
-11000000
-00101000
-01000111
-00101001
-01000010
-10000000
-01101001
-01000001
-11100000
-01000010
-00101001
-01101010
-01000010
-10000000
-01001010
-01100010
-01000101
-10000000
-01001001
-01100001
-01000100
-00011001
