--- output/Bresenhams-2.txt
@@ -1,27 +1,0 @@
-01001101
-01101100
-01000011
-00110001
-01001101
-01101001
-01000100
-10110011
-00111000
-00010001
-10000000
-01100110
-01000110
-00100000
-00110000
-01001100
-01100011
-01000100
-10111010
-01010110
-00100000
-10000000
-01001101
-01100100
-01101001
-01000100
-00011010
