--- output/Bresenhams-3.txt
@@ -1,30 +1,0 @@
-00110001
-10110001
-00111000
-00010001
-10000000
-01100110
-01000110
-00100000
-01101111
-00110010
-10000000
-10111101
-01001100
-01100011
-01000100
-00110000
-10100000
-01010110
-00110010
-00100000
-01101111
-10000000
-10000000
-10111101
-01001100
-01100101
-01000100
-00110000
-10100000
-00011000
