--- output/Fibonacci.txt
@@ -1,13 +1,0 @@
-00111001
-00000000
-00111010
-00000001
-00110000
-01001001
-01100010
-01000001
-10000000
-01001001
-01100010
-01000010
-10100101
