--- output/Linear.txt
@@ -1,13 +1,0 @@
-00111001
-00000011
-00111010
-00000010
-00111011
-00000000
-00110000
-01001001
-01100010
-01000001
-10000000
-01010011
-10100111
