--- output/MemArray.txt
@@ -1,8 +1,0 @@
-00111111
-00000000
-00110000
-00101000
-01111111
-11000000
-01010111
-10100011
