--- output/Pong-1.txt
@@ -1,29 +1,0 @@
-00111001
-00001000
-00111010
-00001000
-00111011
-00000001
-00111100
-00000001
-00111101
-00001000
-00111000
-00001110
-11000000
-00111000
-00001111
-11000001
-00111000
-00000001
-11000010
-01001101
-10000000
-00100000
-01001001
-10000000
-00100000
-01001010
-10000000
-00100000
-00011001
