--- output/Pong-2.txt
@@ -1,32 +1,0 @@
-00101000
-01100101
-01000101
-00100000
-01001011
-01100001
-01000001
-00100000
-01001100
-01100010
-01000010
-00100000
-01001001
-00110010
-10000000
-10111111
-11100000
-01101001
-10000000
-01001010
-10111110
-10000000
-10111111
-11100001
-01101010
-10000000
-10000000
-10111111
-00110000
-10100000
-00011010
-00011011
