--- output/Pong-3.txt
@@ -1,32 +1,0 @@
-01001101
-01101010
-01000110
-00110001
-00111111
-00000111
-10111010
-11100010
-01101110
-00110001
-10110100
-01001011
-01000110
-00110000
-00111111
-00010001
-10111010
-01001110
-01000011
-00011011
-00011000
-10000000
-10000000
-10000000
-10000000
-10000000
-01001000
-01101110
-01000110
-00110000
-01111111
-10100000
