--- output/Pong-4.txt
@@ -1,31 +1,0 @@
-01001001
-00110010
-10000000
-10100110
-00110000
-10101010
-01001000
-01101011
-01000011
-10000000
-01001010
-00110010
-10000000
-10110000
-00110000
-10110100
-01001000
-01101100
-01000100
-10000000
-11100001
-01101010
-00110010
-10000000
-10111011
-00110000
-10111110
-01001000
-01101100
-01000100
-00011001
