--- output/Processes.txt
@@ -1,6 +1,0 @@
-00101000
-01000110
-10000000
-01010111
-01111110
-00011000
