--- output/Sorting.txt
@@ -1,32 +1,0 @@
-00111110
-00000000
-00111111
-00000001
-11100000
-01000001
-10000000
-01111111
-11100000
-01000010
-00110001
-01101001
-10000000
-10110010
-00110000
-01001010
-01000001
-10111000
-01001001
-01111111
-11000000
-01001010
-01111110
-11000000
-01010111
-01101101
-00110010
-10000000
-10100000
-01010110
-00110000
-10100111
