+++ qsymbol_table.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 class symbolizer:
     def __init__(self):
         self.symbol_table = {
