--- tests/Bresenhams-1.s
@@ -1,28 +1,0 @@
-PLD 0
-RST $6
-PST 0
-PLD 1
-RST $1
-PLD 1
-MST 0
-PLD 0
-RST $7
-PLD 1
-RST $2
-NOP
-SUB $1
-RST $1
-MLD 0
-RST $2
-PLD 1
-SUB $2
-RST $2
-NOP
-AST $2
-ADD $2
-RST $5
-NOP
-AST $1
-ADD $1
-RST $4
-SLP 1
