--- tests/Bresenhams-2.s
@@ -1,27 +1,0 @@
-AST $5
-SUB $4
-RST $3
-CND 1
-AST $5
-SUB $1
-RST $4
-JMP 19
-LIM $0
-17
-NOP
-ADD $6
-RST $6
-PST 0
-CND 0
-AST $4
-ADD $3
-RST $4
-JMP 26
-INC $6
-PST 0
-NOP
-AST $5
-ADD $4
-SUB $1
-RST $4
-SLP 2
