--- tests/Bresenhams-3.s
@@ -1,30 +1,0 @@
-CND 1
-JMP 17
-LIM $0
-17
-NOP
-ADD $6
-RST $6
-PST 0
-SUB $7
-CND 2
-NOP
-JMP 29
-AST $4
-ADD $3
-RST $4
-CND 0
-JMP 0
-INC $6
-CND 2
-PST 0
-SUB $7
-NOP
-NOP
-JMP 29
-AST $4
-ADD $5
-RST $4
-CND 0
-JMP 0
-SLP 0
