--- tests/Fibonacci.s
@@ -1,13 +1,0 @@
-LIM $1
-0
-LIM $2
-1
-CND 0
-AST $1
-ADD $2
-RST $1
-NOP
-AST $1
-ADD $2
-RST $2
-JMP 5
