--- tests/Pong-2.s
@@ -1,32 +1,0 @@
-PLD 0
-ADD $5
-RST $5
-PST 0
-AST $3
-ADD $1
-RST $1
-PST 0
-AST $4
-ADD $2
-RST $2
-PST 0
-AST $1
-CND 2
-NOP
-JMP 31
-MLD 0
-SUB $1
-NOP
-AST $2
-JMP 30
-NOP
-JMP 31
-MLD 1
-SUB $2
-NOP
-NOP
-JMP 31
-CND 0
-JMP 0
-SLP 2
-SLP 3
