--- tests/Pong-3.s
@@ -1,32 +1,0 @@
-AST $5
-SUB $2
-RST $6
-CND 1
-LIM $7
-7
-JMP 26
-MLD 2
-SUB $6
-CND 1
-JMP 20
-AST $3
-RST $6
-CND 0
-LIM $7
-17
-JMP 26
-AST $6
-RST $3
-SLP 3
-SLP 0
-NOP
-NOP
-NOP
-NOP
-NOP
-AST $0
-SUB $6
-RST $6
-CND 0
-POI $7
-JMP 0
