--- tests/Pong-4.s
@@ -1,31 +1,0 @@
-AST $1
-CND 2
-NOP
-JMP 6
-CND 0
-JMP 10
-AST $0
-SUB $3
-RST $3
-NOP
-AST $2
-CND 2
-NOP
-JMP 16
-CND 0
-JMP 20
-AST $0
-SUB $4
-RST $4
-NOP
-MLD 1
-SUB $2
-CND 2
-NOP
-JMP 27
-CND 0
-JMP 30
-AST $0
-SUB $4
-RST $4
-SLP 1
