--- tests/Processes.s
@@ -1,6 +1,0 @@
-PLD 0
-RST $6
-NOP
-INC $7
-POI $6
-SLP 0
