--- tests/Sorting.s
@@ -1,32 +1,0 @@
-LIM $6
-0
-LIM $7
-1
-MLD 0
-RST $1
-NOP
-POI $7
-MLD 0
-RST $2
-CND 1
-SUB $1
-NOP
-JMP 18
-CND 0
-AST $2
-RST $1
-JMP 24
-AST $1
-POI $7
-MST 0
-AST $2
-POI $6
-MST 0
-INC $7
-SUB $5
-CND 2
-NOP
-JMP 0
-INC $6
-CND 0
-JMP 7
